Add files via upload
To prevent cyberattacks in advance, vulnerability analysis in computers and operating systems involves identifying and evaluating any security holes. It is a systematic check of security weakness in an organization information system. This work as a threat assessment and is used to evaluate how susceptible a network may be to future cyberattacks and attempted hacks. It is essential for ensuring security, lowering risk, adhering to regulations, managing patches effectively, organising incident responses, cutting costs, improving continuously, and preventing zero-day exploits. Regular analysis is essential for keeping a solid cybersecurity posture in the ever-changing threat landscape of today. 

Overall, vulnerability analysis is merely one component of the greater field of ethical hacking. Ethical hackers operate legally and purposefully to get access to a company’s networks and systems to evaluate the company’s existing security and defences.

This report aims to provide a thorough overview of vulnerability analysis and its exploitation technique which indicates the significance of the security to get rid from the attacker. This report will go over different vulnerability that arises in Metasploitable2 and Windows 7.